requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
twilio==9.4.4
requests==2.32.3
openai==1.61.1
pydantic==2.10.6
python-multipart